fix(perf): ensure float32 output in create_gaussian_kernel_1d and slic
- create_gaussian_kernel_1d: use float32 throughout (sigma!=0 branch was
  returning float64 due to np.exp promotion); keeps dtype consistent
- slic: cast max_val and 1e-6 to float32 to avoid float64 upcast in
  image_normalized division

Addresses Copilot PR review feedback.

Made-with: Cursor

+++ albumentations/augmentations/pixel/functional.py
@@ -2501,8 +2501,8 @@ def slic(
     num_pixels = height * width
 
     # Normalize image to [0, 1] range
-    max_val = image.max()
-    image_normalized = image.astype(np.float32) / (max_val + 1e-6)
+    max_val = np.float32(image.max())
+    image_normalized = image.astype(np.float32) / (max_val + np.float32(1e-6))
 
     # Initialize cluster centers via meshgrid
     grid_step = int((num_pixels / n_segments) ** 0.5)
